feat: update routing to better match actual reddit's routing
