+++ .gitignore
@@ -83,3 +83,4 @@ yarn-error.log*
 
 # next.js analyze bundle
 /bundles
+.vercel
