+++ now.json
@@ -1,5 +1,4 @@
 {
   "version": 2,
-  "name": "sleddit",
   "alias": "sleddit"
 }
